bio-tools/pyproject.toml
[project]
name = "bio-tools"
version = "0.1.0"
description = "MCP server exposing biological tool APIs"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
